Cleanup #1 for GCC 3.4: (patch by Zachary A. Clifford) * Added typedef genf for function pointers (adopted from MAME) * Put "#ifndef PINMAME" around unused functions * Increased size of pcm_pos and sOut in TMS320AV120Chip struture to fit buffer size * Renamed functions that collide with gcc/libc built-in functions
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@2896 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/tms320av120.c
@@ -101,8 +101,8 @@ struct TMS320AV120Chip
  UINT8  framebuff[MPG_FRAMESIZE];		//Holds raw mpg data for 1 frame (also used as header input buffer)
  UINT8	fb_pos;							//Current frame buffer position
  INT16  pcmbuffer[CAP_PCMBUFFER_SIZE];	//Decoded PCM data buffer
- UINT16 pcm_pos;						//Position of next data to be input into the pcm buffer
- UINT16 sOut;							//Position of next sample to read out of the pcm buffer
+ UINT32 pcm_pos;						//Position of next data to be input into the pcm buffer
+ UINT32 sOut;							//Position of next sample to read out of the pcm buffer
  int    stream;							//Holds stream channel assignment
  int    bitsRemaining;					//Keep track of # of bits we've read from frame buffer
  long *_V[16];							//Synthesis window for single channel
